+++ RELEASES.md
@@ -521,3 +521,4 @@
 8.1.2 2018-11-07
 ----------------
   - fix: issue with __transformResponse protocol happn_1, bad callback
+  - fix: security, users that are substrings of other users get returned by getUser